Refactor handlers into multiple files
