+++ client/deployment_test.go
